+++ mapstruct-mapper-repo/generator/src/main/java/org/mapstruct/example/repo/Mapper.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.cat.map;
+package org.mapstruct.example.repo;
 
 import java.util.List;
